Fix warnings when compiling with G++ 4.1.1 on Fedora Core 6
git-svn-id: svn://svn.minisip.org/minisip/trunk@3274 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmcrypto/source/openssl/cert.cxx
@@ -420,7 +420,7 @@ int ossl_certificate::get_der_length(){
 }
 
 void ossl_certificate::get_der( unsigned char * output, unsigned int * length ){
-	if( *length < get_der_length() ){
+	if( *length < (unsigned int)get_der_length() ){
  		throw certificate_exception(
 			"Given buffer is to short" );
 	}
